+++ firmware/targets/f7/target.json
@@ -45,7 +45,6 @@
         "flipper_application",
         "flipperformat",
         "toolbox",
-        "flipper7",
-        "nfc"
+        "flipper7"
     ]
 }
